Move selectOption into common.go for reused driver tools
